requirements.txt
streamlit>=1.31.0
openai>=1.12.0
yfinance>=0.2.36
python-dotenv>=1.0.0
